CXP-426: Fix naming
